% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binary_cross_entropy}
\alias{binary_cross_entropy}
\title{Binary cross-entropy loss}
\usage{
binary_cross_entropy(ypred, ytrue)
}
\arguments{
\item{ypred}{vector of predicted probabilities}

\item{ytrue}{vector of true labels {0, 1}}
}
\description{
Computes the binary cross-entropy loss (log-loss) of two vectors
}
