% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_regression-fit.R
\name{cox_regression}
\alias{cox_regression}
\alias{cox_regression.default}
\alias{cox_regression.data.frame}
\alias{cox_regression.matrix}
\alias{cox_regression.formula}
\alias{cox_regression.recipe}
\title{Fit a \code{cox_regression}}
\usage{
cox_regression(x, ...)

\method{cox_regression}{default}(x, ...)

\method{cox_regression}{data.frame}(x, y, ...)

\method{cox_regression}{matrix}(x, y, ...)

\method{cox_regression}{formula}(formula, data, ...)

\method{cox_regression}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{cox_regression} object.
}
\description{
\code{cox_regression()} fits a model.
}
\examples{
set.seed(0)
data <- simulate_dummy_surv_data()

model <- cox_regression(event~time+., data, l1=0.05, l2=0.01, frob=0.01, num_comp=3)
model

}
