% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{error_summary_CEN_surv}
\alias{error_summary_CEN_surv}
\title{Evaluate survival simulations}
\usage{
error_summary_CEN_surv(model, new_data)
}
\arguments{
\item{model}{A \code{linear_classification} model}

\item{new_data}{Data generated from \code{simulate_CEN_surv_data}}
}
\description{
Evaluate survival simulations
}
