% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasp-package.R
\docType{package}
\name{glasp-package}
\alias{glasp}
\alias{glasp-package}
\title{glasp: Group Linear Algorithm with Sparse Principal decomposition}
\description{
Implements the the Group Linear Algorithm with Sparse Principal decomposition, an algorithm for supervised variable selection and clustering. Our approach extends the Sparse-Group Lasso regularization to calculate clusters as part of the model fit. Therefore, unlike Sparse-Group Lasso, our idea does not require prior specification of clusters between variables. To determine the clusters, we solve a particular case of sparse Singular Value Decomposition, with a regularization term that follows naturally from the Group Lasso penalty. Moreover, this paper proposes a unified implementation to deal with, but not limited to, linear regression, logistic regression, and proportional hazards models with right-censoring.
}
\author{
\strong{Maintainer}: Juan C. Laria \email{juank.laria@gmail.com} (\href{https://orcid.org/0000-0001-7734-9647}{ORCID})

Other contributors:
\itemize{
  \item Rosa E. Lillo [thesis advisor]
  \item M. Carmen Aquilera-Morillo [thesis advisor]
}

}
\keyword{internal}
