% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasp_cox.R
\name{glasp_cox}
\alias{glasp_cox}
\title{parsnip interface for GLASP - Cox Models}
\usage{
glasp_cox(
  mode = "classification",
  l1 = NULL,
  l2 = NULL,
  frob = NULL,
  num_comp = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.}

\item{l1}{Regularization in norm l1 (lasso)}

\item{l2}{Regularization in norm l2 (ridge)}

\item{frob}{Regularization in Frobenius norm. It is unique to glasp and controls the importance
of the clustering on the variables. If frob=0, we want to fit a model equivalent to elastic-net,
without clustering. If frob > l1+l2 means that we are more interested on the feature clustering
than the variable selection. If frob > 1 we are more interested on finding clusters than on
getting a linear model with good predictive power.}

\item{num_comp}{Maximum number of clusters to search.}
}
\value{
A \code{glasp_cox} parsnip model
}
\description{
parsnip interface for GLASP - Cox Models
}
