% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasp_model.R
\name{glasp_model}
\alias{glasp_model}
\title{parsnip interface for Variable Cluster - Linear Regression Models}
\usage{
glasp_model(
  mode = "unknown",
  l1 = NULL,
  l2 = NULL,
  frob = NULL,
  num_comp = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.}

\item{l1}{An non-negative number representing the total amount of regularization L1}

\item{l2}{An non-negative number representing the total amount of group regularization L2}

\item{frob}{An non-negative number representing the total amount of regularization in the Frobenius norm of the error of the low-rank approximation}

\item{num_comp}{Number of clusters to search for}
}
\description{
parsnip interface for Variable Cluster - Linear Regression Models
}
\examples{
\dontrun{
library(glasp)
set.seed(0)
# generate data -----------------------------------------------------------
data <- glasp::simulate_dummy_surv_data(N = 100)
# compute glasp -----------------------------------------------------------
glasp_mod <- glasp::glasp_model(l1 = tune::tune(),
                                l2 = tune::tune(),
                                frob = tune::tune(),
                                num_comp = 3)

glasp_mod <- parsnip::set_mode(glasp_mod, "classification")
glasp_mod <- parsnip::set_engine(glasp_mod, "glasp")

data_rs <- rsample::vfold_cv(data, v = 10)
metric_vals <- yardstick::metric_set(yardstick::roc_auc)
ctrl <- tune::control_grid(verbose = T)

rec_form <- tune::tune_grid(glasp_mod, event~.,
                            resamples = data_rs,
                            metrics = metric_vals,
                            control = ctrl,
                            grid = 100)

best_config <- show_best(rec_form, metric = "yardstick::roc_auc")
model <- glasp::linear_classification( event~., data,
                                       l1 = best_config$l1[1],
                                       l2 = best_config$l2[1],
                                       frob = best_config$frob[1],
                                       num_comp = 3)

model$beta
}
}
