% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_classification-fit.R
\name{linear_classification}
\alias{linear_classification}
\alias{linear_classification.default}
\alias{linear_classification.data.frame}
\alias{linear_classification.matrix}
\alias{linear_classification.formula}
\alias{linear_classification.recipe}
\title{Fit a \code{linear_classification}}
\usage{
linear_classification(x, ...)

\method{linear_classification}{default}(x, ...)

\method{linear_classification}{data.frame}(x, y, ...)

\method{linear_classification}{matrix}(x, y, ...)

\method{linear_classification}{formula}(formula, data, ...)

\method{linear_classification}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{linear_classification} object.
}
\description{
\code{linear_classification()} fits a model.
}
