% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression-fit.R
\name{logistic_regression}
\alias{logistic_regression}
\alias{logistic_regression.default}
\alias{logistic_regression.data.frame}
\alias{logistic_regression.matrix}
\alias{logistic_regression.formula}
\alias{logistic_regression.recipe}
\title{Fit a \code{logistic_regression}}
\usage{
logistic_regression(x, ...)

\method{logistic_regression}{default}(x, ...)

\method{logistic_regression}{data.frame}(x, y, ...)

\method{logistic_regression}{matrix}(x, y, ...)

\method{logistic_regression}{formula}(formula, data, ...)

\method{logistic_regression}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{logistic_regression} object.
}
\description{
\code{logistic_regression()} fits a model.
}
\examples{
set.seed(0)
data <- simulate_dummy_logistic_data()

model <- logistic_regression(y~., data, l1=0.05, l2=0.01, frob=0.01, num_comp=3)
model

new_data <- simulate_dummy_logistic_data()

predict(model, new_data, type = "numeric")
predict(model, new_data, type = "prob")
predict(model, new_data, type = "class")
}
