% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_squared_error}
\alias{mean_squared_error}
\title{Mean squared error}
\usage{
mean_squared_error(ypred, ytrue)
}
\arguments{
\item{ypred}{vector of predictions}

\item{ytrue}{vector of true labels}
}
\description{
Computes the mean squared error of two vectors
}
