% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_CEN_surv_data}
\alias{simulate_CEN_surv_data}
\title{Simulate survival data with groups}
\usage{
simulate_CEN_surv_data(
  N = 100,
  p = 30,
  p_group = 10,
  true_p_group = 3,
  rho = 0.5
)
}
\arguments{
\item{N}{Number of observations (event or censored time)}

\item{p}{Number of features}

\item{p_group}{Number of features in each group}

\item{true_p_group}{Number of important features in each group}

\item{rho}{Correlation inside groups}
}
\value{
A data set0   2
}
\description{
Simulate survival data with groups
}
