% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_dummy_logistic_data}
\alias{simulate_dummy_logistic_data}
\title{Simulate toy logistic data for benchmark}
\usage{
simulate_dummy_logistic_data(N = 100, p = 10, nzc = 5)
}
\arguments{
\item{N}{Number of observations}

\item{p}{Number of features}

\item{nzc}{True number of features in the model}
}
\value{
An expanded data set
}
\description{
Simulate toy logistic data for benchmark
}
