% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_dummy_surv_data}
\alias{simulate_dummy_surv_data}
\title{Simulate toy survival data for benchmark}
\usage{
simulate_dummy_surv_data(N = 100, p = 10, nzc = 5, censProb = 0.3)
}
\arguments{
\item{N}{Number of observations (event or censored time)}

\item{p}{Number of features}

\item{nzc}{True number of features in the model}

\item{censProb}{Probability of right censoring}
}
\value{
An expanded data set
}
\description{
Simulate toy survival data for benchmark
}
