% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_dist.R
\name{vi_dist}
\alias{vi_dist}
\title{Variation of information distance}
\usage{
vi_dist(cl1, cl2, parts = FALSE, base = 2)
}
\arguments{
\item{cl1}{Vector of cluster membership indices}

\item{cl2}{Vector of cluster membership indices}

\item{parts}{logical; should the two conditional entropies also be returned?}

\item{base}{base of logarithm used for computation of entropy and mutual information}
}
\value{
The VI distance. If parts=TRUE the two conditional entropies are appended.
}
\description{
(From 'mcclust' R package) Computes the 'variation of information'
distance of Meila (2007) between two clusterings/partitions of the same objects.
}
\details{
The variation of information distance is the sum of the two conditional
entropies of one clustering given the other. For details see Meila (2007).
}
\references{
Meila, M. (2007) Comparing Clusterings - an Information Based Distance. Journal of Multivariate Analysis, 98, 873 – 895.
}
