% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression-predict.R
\name{predict.linear_regression}
\alias{predict.linear_regression}
\title{Predict from a \code{linear_regression}}
\usage{
\method{predict}{linear_regression}(object, new_data, type = "numeric", ...)
}
\arguments{
\item{object}{A \code{linear_regression} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"numeric"} for numeric predictions.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{linear_regression}
}
\examples{
set.seed(0)
data <- simulate_dummy_linear_data()

model <- linear_regression(y~., data, l1=0.05, l2=0.01, frob=0.01, num_comp=3)

new_data <- simulate_dummy_linear_data()

predict(model, new_data, type = "numeric")

}
